+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Next version
 
+- Add defaults for `minimum{Base,Quote}PerOfferFactor` in the Kandel configuration
+- Add initial configuration for Blast and Blast Sepolia. For now, they use the defaults, so have only be added to make it easy to find and modify later if needed.
+
 # 2.0.5-14
 
 - fix: Handle missing case where AaveKandel was assumed available