+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+# 2.0.5-16
+
 - Add mangrove amplifier as export on `index.ts`
 
 # 2.0.5-15